requirements.txt
schedule
python-dotenv
# bot
python-telegram-bot==13.15
# scraping
requests
beautifulsoup4